% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_daily_map.R
\name{covid_daily_map}
\alias{covid_daily_map}
\title{coronavirus cases shown on the globe}
\usage{
covid_daily_map(date = "20-10-23", type = "deaths")
}
\arguments{
\item{date}{the map will be from data of the date you specify.
put in "yy-mm-dd" format. date can be from "20-01-22" till now.}

\item{type}{type of cases you're interested in. can be in c("confirmed",
"deaths", "recovered").}
}
\value{
a ggplot of world map containing coronavirus cases.
}
\description{
If function get_covid_date() did not run first, it runs it and
then draws the plot.
}
\examples{
\dontrun{
covid_daily_map(date == "20-09-28", type == "confirmed")
}
}
\seealso{
\code{\link[=get_covid_data]{get_covid_data()}}
}
