% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_time_series.R
\name{covid_time_series}
\alias{covid_time_series}
\title{time series of coronavirus cases in a given interval}
\usage{
covid_time_series(
  country = "Iran",
  interval_start = "20-05-22",
  interval_end = Sys.Date()
)
}
\arguments{
\item{country}{country you're interested in its covid19 cases. Note that
country input is case sensitive, for example "Iran". USA is present as "US" and
UK is accessible as "United Kingdom".}

\item{interval_start}{start date of interval. Put in "yy-mm-dd" format.
Can be from "20-01-22" till now.}

\item{interval_end}{(optional) end date of interval. if not specified
defaults to today.}
}
\value{
a barplot of covid19 cases
}
\description{
this function takes country and time interval, and draws a plot
of confirmed cases, recovered people, and deaths in that interval.
}
\examples{
\dontrun{
covid_time_series(country = "Iran", interval_star = "20-03-26")
covid_time_series(country = "Iran", interval_star = "20-03-26", interval_end = "20-06-22")
}
}
