% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covid_data.R
\name{get_covid_data}
\alias{get_covid_data}
\title{Get Covid-19 Data}
\usage{
get_covid_data()
}
\value{
a data.frame of updated covid19 data
}
\description{
Gets data of confirmed cases, deaths, and recovered cases of
covid-19 around the globe. columns contain type ( c("confirmed","deaths","recovered") ),
country, lat, long, and days (start from 2020-01-22 until now)
}
\examples{
get_covid_date()
}
\references{
\url{https://github.com/CSSEGISandData/COVID-19}
}
